pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testforge"
version = "0.1.0"
description = "LLM-powered QA automation platform — from documents to test reports"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "LucaBlaq Tech", email = "songblaq@gmail.com" },
]
keywords = ["testing", "qa", "automation", "llm", "playwright"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.1",
    "playwright>=1.40",
    "httpx>=0.25",
    "PyMuPDF>=1.23",
    "python-pptx>=0.6",
    "openpyxl>=3.1",
    "python-docx>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
all = ["testforge[anthropic,openai]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
testforge = "testforge.cli:cli"

[project.urls]
Homepage = "https://github.com/songblaq/testforge"
Repository = "https://github.com/songblaq/testforge"
Issues = "https://github.com/songblaq/testforge/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
